Backend/requirements.txt
greenlet==3.3.1
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.46
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
xgboost==1.7.5
apscheduler==3.10.4